refactor: organize chart services into dedicated chart-tools folder
Moved chart transformation services into a dedicated folder structure:
- chart-ai-tools.ts → chart-tools/types.ts (clearer naming)
- graph-transformer.ts → chart-tools/ai-transformer.ts (descriptive name)
- Updated import references in dashboard router

This improves code organization and makes the purpose of each file more explicit.
